requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
dataclasses==0.6
greenlet==3.0.3
idna==3.7
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pip-system-certs==4.0
PyMySQL==1.1.1
python-certifi-win32==1.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
pytz==2024.1
requests==2.32.3
sendgrid==6.11.0
setuptools-scm==8.1.0
six==1.16.0
SQLAlchemy==2.0.31
starkbank-ecdsa==2.2.0
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0